refactor: extract shared memory repository class
