+++ src/resources/boards/board.memory.repository.js
@@ -1,0 +1,6 @@
+import { MemoryRepository } from '../../common/memory-repository.js';
+
+/**
+ * @type { MemoryRepository<Board> }
+ */
+export const boardsRepo = new MemoryRepository();
